Added in music Ocean manipulation
The music can now be reloaded at runtime

The music:

-Now makes the waves jump up and down
-Deterimines the frequency of the waves

+++ Assets/Song.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Song : MonoBehaviour
 {
-    public int m_BeatInterval;
+    public float m_BeatInterval;
     [SerializeField] public AudioClip m_Clip;
 
     public string m_SongName;
